backend/faust_worker/requirements.txt
faust-streaming==0.10.4
aiohttp==3.8.4
pydantic==1.10.6